requirements.txt
# Зависимости для нового AI процессора психологических статей

# Основные библиотеки
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
python-dotenv>=1.0.0

# Дополнительные библиотеки
lxml>=4.9.0
html5lib>=1.1

# Для работы с изображениями (опционально)
Pillow>=10.0.0

# Для логирования и мониторинга
colorlog>=6.7.0

# Для работы с датами
python-dateutil>=2.8.0